Combining multiple commits into 1 which contains a series of bug fixes for both linux & windows.
Fixed null reference bug in opencl benchmark if timer was not found
Added an event wait for unmapping a memory region

Fixed copy and paste bug in cpu coo2csr, which caused a junk
buffer in linux for the csr data.

Fixing compiler error on windows; the windows headers already
define FLOAT/DOUBLE.

Fixing more access violations in the cl benchmark client when
the timing library is not present.