Add cmake option to choose between system and built cl2.hpp
`USE_SYSTEM_CL2HPP` cmake option has been added to the project
build options. This is a boolean variable, which is set to `OFF`
by default. The default behavior is to clone the `OpenCL-CLHPP` repository
and generate the cl2.hpp header. When the cmake option `USE_SYSTEM_CL2HPP`
is set to `ON` the cl2.hpp header is expected to be available in system
paths so that the build process automatically picks it up.