Replace cl.hpp with cl2.hpp (#187)
* Add missing cpp header inclusion for sqrt function

* Replace cl.hpp header with cl2.hpp

* Added OpenCL-CLHPP repository as external project for library and samples
  projects.
* Modified all source files of library and samples to reflect the usage of
  cl2.hpp header.

* Add cmake option to choose between system and built cl2.hpp

`USE_SYSTEM_CL2HPP` cmake option has been added to the project
build options. This is a boolean variable, which is set to `OFF`
by default. The default behavior is to clone the `OpenCL-CLHPP` repository
and generate the cl2.hpp header. When the cmake option `USE_SYSTEM_CL2HPP`
is set to `ON` the cl2.hpp header is expected to be available in system
paths so that the build process automatically picks it up.

* Remove deprecated add_compiler_export_flags cmake fn call

* Fix cmake syntax in cl2.hpp conditional statement

* Use set_target_properties cmake fn to set `-fvisibility` flag

* Change travis ci osx version to 10.11/xcode8

* Add command to install cmake on osx platform

`osx_image: xcode8` image of travis ci doesn't come with cmake
preinstalled althought earlier versions do. Hence, added the cmake
installation command to before install stage.